README.md
This is my first pic 10c hw assignment. The task was to create a Spanish card game similar to that of the classic casino game 21. We were also meant to become familiar with git and using a repository to save and commit code to. This was an introduction to version control with the use of multiple branches.